init modules for core components; bugfix for source parsers
