+++ server/core/article_parser/__init__.py
@@ -1,0 +1,1 @@
+from .article_parser import ArticleParser
