+++ server/core/fetcher/__init__.py
@@ -1,0 +1,1 @@
+from .fetcher import Fetcher
