+++ server/core/incident_predictor/__init__.py
@@ -1,0 +1,1 @@
+from .incident_predictor import IncidentPredictor
