+++ server/core/source_parser/__init__.py
@@ -1,0 +1,1 @@
+from .source_parser import SourceParser
