+++ server/core/source_parser/parsers/vk_parser.py
@@ -1,9 +1,9 @@
 from typing import Iterable
 
 import re
-from lxml import etree
+from lxml import cssselect, etree
 
-from server.core.article_parser.parsers.base_parser import ParserBase
+from .base_parser import ParserBase
 
 
 class VkParser(ParserBase):
@@ -13,7 +13,7 @@ class VkParser(ParserBase):
 
     @classmethod
     def extract_urls(cls, url: str, document=None) -> Iterable[str]:
-        links = CSSSelector("a.PostHeaderSubtitle__link")(document)
+        links = cssselect.CSSSelector("a.PostHeaderSubtitle__link")(document)
 
         for link in links:
             news_page_link = "https://vk.com" + link.get("href")
